relocate chapter data class
